+++ src/Handler/StreamHandler.php
@@ -454,6 +454,13 @@ class StreamHandler
     private static function triggerUnsupportedRequestOptionDeprecations(RequestInterface $request, array $options): void
     {
         if (
+            \array_key_exists('curl_share', $options)
+            && CurlShareHandleState::normalizeMode($options['curl_share'], 'curl_share') !== CurlShare::NONE
+        ) {
+            \trigger_deprecation('guzzlehttp/guzzle', '7.11', 'Passing the "curl_share" option to the stream handler is deprecated; guzzlehttp/guzzle 8.0 will reject this option because the stream handler does not support cURL sharing.');
+        }
+
+        if (
             \array_key_exists('curl', $options)
             && $options['curl'] !== null
             && $options['curl'] !== []