+++ tests/bootstrap-phpstan.php
@@ -7,3 +7,9 @@ if (!\defined('IDNA_DEFAULT')) {
 if (!\defined('INTL_IDNA_VARIANT_UTS46')) {
     \define('INTL_IDNA_VARIANT_UTS46', 1);
 }
+
+if (!\class_exists('CurlShareHandle')) {
+    class CurlShareHandle
+    {
+    }
+}